+++ src/main/java/com/dev/alex/Repository/HoldingsRepository.java
@@ -7,6 +7,7 @@ import org.springframework.data.mongodb.repository.Update;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Repository
@@ -19,7 +20,7 @@ public interface HoldingsRepository extends MongoRepository<Holdings, String> {
    @Query(value = "{ 'holdingId': ?0 }")
    @Update("{ '$set': { 'averagePurchasePrice': :#{#averagePurchasePrice}, 'quantity': :#{#quantity} }, '$currentDate': { 'updatedAt': true } }")
    void updateAveragePurchasePriceAndQuantity(@Param("holdingId") String holdingId,
-                                              @Param("averagePurchasePrice") Double averagePurchasePrice,
-                                              @Param("quantity") Double quantity
+                                              @Param("averagePurchasePrice") BigDecimal averagePurchasePrice,
+                                              @Param("quantity") BigDecimal quantity
    );
 }
